# openapi.yaml
openapi: 3.0.3
info:
  title: GameShop REST API
  version: 1.0.0
  description: |
    REST API курсового проекту GameShop. Покриває Лаби 5-6: реєстрація/авторизація,
    відгуки, замовлення, проксі-каталог до RAWG.io, адмінські операції.
servers:
  - url: http://localhost:3000
    description: Локальний dev-сервер

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gs_sid
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        role: { type: string, enum: [user, admin, seller] }
        avatar: { type: string, nullable: true }
    Product:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        released: { type: string, nullable: true }
        cover: { type: string, nullable: true }
        rating: { type: number }
        metacritic: { type: integer, nullable: true }
        genres: { type: array, items: { type: string } }
        platforms: { type: array, items: { type: string } }
        price: { type: integer }
    ProductList:
      type: object
      properties:
        count: { type: integer }
        hasNext: { type: boolean }
        results: { type: array, items: { $ref: '#/components/schemas/Product' } }
    Review:
      type: object
      properties:
        id: { type: integer }
        product_id: { type: string }
        user_id: { type: integer, nullable: true }
        author: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        body: { type: string }
        created_at: { type: string }
    ReviewPage:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        hasNext: { type: boolean }
        items: { type: array, items: { $ref: '#/components/schemas/Review' } }
    OrderInput:
      type: object
      required: [customer, items]
      properties:
        customer:
          type: object
          required: [fullName, email, phone, city, address]
          properties:
            fullName: { type: string }
            email: { type: string }
            phone: { type: string }
            city: { type: string }
            address: { type: string }
            comment: { type: string }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [id, title, price, qty]
            properties:
              id: { type: string }
              title: { type: string }
              price: { type: integer }
              qty: { type: integer }
    Order:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer, nullable: true }
        full_name: { type: string }
        email: { type: string }
        phone: { type: string }
        city: { type: string }
        address: { type: string }
        comment: { type: string, nullable: true }
        status: { type: string, enum: [new, paid, shipped, cancelled] }
        total: { type: integer }
        created_at: { type: string }
    Error:
      type: object
      properties:
        error: { type: string }

paths:
  /api/products:
    get:
      summary: Список товарів (проксі до RAWG + локальні ціни)
      parameters:
        - { in: query, name: search, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, default: 12 } }
        - { in: query, name: ordering, schema: { type: string, default: '-rating' } }
        - { in: query, name: genre, schema: { type: string } }
        - { in: query, name: platform, schema: { type: string } }
      responses:
        '200':
          description: Сторінка товарів
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductList' }
  /api/products/{slug}:
    get:
      summary: Деталі товару
      parameters:
        - { in: path, name: slug, required: true, schema: { type: string } }
      responses:
        '200': { description: Товар, content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } } }
        '404': { description: Не знайдено, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  /api/reviews/{productId}:
    get:
      summary: Список відгуків (порціями)
      parameters:
        - { in: path, name: productId, required: true, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, default: 5 } }
      responses:
        '200': { description: Сторінка відгуків, content: { application/json: { schema: { $ref: '#/components/schemas/ReviewPage' } } } }
    post:
      summary: Додати відгук (потрібна авторизація)
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: productId, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating, body]
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                body: { type: string, minLength: 5, maxLength: 2000 }
      responses:
        '201': { description: Створений відгук, content: { application/json: { schema: { $ref: '#/components/schemas/Review' } } } }
        '401': { description: Не авторизовано }

  /api/auth/register:
    post:
      summary: Реєстрація
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username: { type: string }
                email: { type: string }
                password: { type: string, minLength: 6 }
      responses:
        '201': { description: Створено, content: { application/json: { schema: { type: object, properties: { user: { $ref: '#/components/schemas/User' } } } } } }
        '409': { description: Користувач уже існує }

  /api/auth/login:
    post:
      summary: Авторизація
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200': { description: Успіх, content: { application/json: { schema: { type: object, properties: { user: { $ref: '#/components/schemas/User' } } } } } }
        '401': { description: Невірний логін/пароль }

  /api/auth/logout:
    post:
      summary: Вихід
      responses:
        '200': { description: OK }

  /api/auth/me:
    get:
      summary: Поточний користувач
      responses:
        '200':
          description: User або null
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    nullable: true
                    allOf: [{ $ref: '#/components/schemas/User' }]

  /api/orders:
    get:
      summary: Замовлення поточного користувача
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: Список, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Order' } } } } }
    post:
      summary: Створити замовлення
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderInput' }
      responses:
        '201': { description: Створено, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }

  /api/upload/avatar:
    post:
      summary: Завантажити аватар (multipart/form-data)
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: URL аватара
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }

  /api/admin/users:
    get:
      summary: '[admin] Список користувачів'
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: OK }

  /api/admin/users/{id}:
    patch:
      summary: '[admin] Заблокувати/розблокувати'
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        '200': { description: OK }

  /api/admin/prices:
    get:
      summary: '[admin] Усі локальні ціни'
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: OK }

  /api/admin/prices/{slug}:
    put:
      summary: '[admin] Встановити ціну'
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: slug, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [price]
              properties:
                price: { type: integer }
      responses:
        '200': { description: OK }
    delete:
      summary: '[admin] Видалити локальну ціну'
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: slug, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }

  /api/admin/orders:
    get:
      summary: '[admin] Усі замовлення'
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: OK }
  /api/admin/orders/{id}:
    patch:
      summary: '[admin] Змінити статус замовлення'
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [new, paid, shipped, cancelled] }
      responses:
        '200': { description: OK }